backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.9
blinker==1.9.0
certifi==2025.4.26
click==8.2.1
colorama==0.4.6
dill==0.4.0
distro==1.9.0
Flask==3.1.1
flask-cors==6.0.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
MarkupSafe==3.0.2
mccabe==0.7.0
numpy==2.2.6
openai==1.82.0
packaging==25.0
platformdirs==4.3.7
pydantic==2.11.5
pydantic_core==2.33.2
pylint==3.3.6
python-dotenv==1.1.0
scikit-learn==1.6.1
scipy==1.15.3
sniffio==1.3.1
threadpoolctl==3.6.0
tomlkit==0.13.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
Werkzeug==3.1.3
